+++ common/src/main/java/com/alessiodp/parties/common/commands/sub/CommandLeave.java
@@ -9,21 +9,34 @@ import com.alessiodp.parties.api.events.common.party.IPartyPreDeleteEvent;
 import com.alessiodp.parties.api.events.common.player.IPlayerPostLeaveEvent;
 import com.alessiodp.parties.api.events.common.player.IPlayerPreLeaveEvent;
 import com.alessiodp.parties.common.PartiesPlugin;
+import com.alessiodp.parties.common.commands.list.CommonCommands;
 import com.alessiodp.parties.common.commands.utils.PartiesCommandData;
 import com.alessiodp.parties.common.commands.utils.PartiesSubCommand;
 import com.alessiodp.parties.common.configuration.PartiesConstants;
+import com.alessiodp.parties.common.configuration.data.ConfigMain;
+import com.alessiodp.parties.common.configuration.data.ConfigParties;
 import com.alessiodp.parties.common.configuration.data.Messages;
 import com.alessiodp.parties.common.parties.objects.PartyImpl;
-import com.alessiodp.parties.common.commands.utils.PartiesPermission;
+import com.alessiodp.parties.common.utils.PartiesPermission;
 import com.alessiodp.parties.common.players.objects.PartyPlayerImpl;
 import com.alessiodp.parties.api.enums.DeleteCause;
-import lombok.Getter;
 
 public class CommandLeave extends PartiesSubCommand {
-	@Getter private final boolean executableByConsole = false;
 	
 	public CommandLeave(ADPPlugin plugin, ADPMainCommand mainCommand) {
-		super(plugin, mainCommand);
+		super(
+				plugin,
+				mainCommand,
+				CommonCommands.LEAVE,
+				PartiesPermission.USER_LEAVE,
+				ConfigMain.COMMANDS_CMD_LEAVE,
+				false
+		);
+		
+		syntax = baseSyntax();
+		
+		description = Messages.HELP_MAIN_DESCRIPTIONS_LEAVE;
+		help = Messages.HELP_MAIN_COMMANDS_LEAVE;
 	}
 	
 	@Override
@@ -32,8 +45,8 @@ public class CommandLeave extends PartiesSubCommand {
 		PartyPlayerImpl partyPlayer = ((PartiesPlugin) plugin).getPlayerManager().getPlayer(sender.getUUID());
 		
 		// Checks for command prerequisites
-		if (!sender.hasPermission(PartiesPermission.LEAVE.toString())) {
-			sendNoPermissionMessage(partyPlayer, PartiesPermission.LEAVE);
+		if (!sender.hasPermission(permission)) {
+			sendNoPermissionMessage(partyPlayer, permission);
 			return false;
 		}
 		
@@ -63,38 +76,58 @@ public class CommandLeave extends PartiesSubCommand {
 		if (!partiesPreLeaveEvent.isCancelled()) {
 			if (party.getLeader() != null && party.getLeader().equals(sender.getUUID())) {
 				// Is leader
-				// Calling Pre API event
-				IPartyPreDeleteEvent partiesPreDeleteEvent = ((PartiesPlugin) plugin).getEventManager().preparePartyPreDeleteEvent(party, DeleteCause.LEAVE, null, partyPlayer);
-				((PartiesPlugin) plugin).getEventManager().callEvent(partiesPreDeleteEvent);
+				boolean mustDelete = true;
+				// Check if leader can be changed
+				if (ConfigParties.GENERAL_MEMBERS_CHANGE_LEADER_ON_LEAVE
+						&& party.getMembers().size() > 1) {
+					PartyPlayerImpl newLeader = party.findNewLeader();
+					
+					if (newLeader != null) {
+						// Found a new leader
+						mustDelete = false;
+						
+						party.changeLeader(newLeader);
+						party.removeMember(partyPlayer);
+						
+						sendMessage(sender, partyPlayer, Messages.MAINCMD_LEAVE_LEFT, party);
+						party.broadcastMessage(Messages.MAINCMD_LEAVE_LEADER_CHANGED, newLeader);
+						
+						plugin.getLoggerManager().logDebug(String.format(PartiesConstants.DEBUG_CMD_LEAVE_LEADER_CHANGE,
+								partyPlayer.getName(), party.getName(), newLeader.getName()), true);
+					}
+				}
 				
-				if (!partiesPreDeleteEvent.isCancelled()) {
-					// Disbanding party
-					sendMessage(sender, partyPlayer, Messages.MAINCMD_LEAVE_LEFT, party);
-					party.broadcastMessage(Messages.MAINCMD_LEAVE_DISBANDED, partyPlayer);
+				if (mustDelete) {
+					// Calling Pre API event
+					IPartyPreDeleteEvent partiesPreDeleteEvent = ((PartiesPlugin) plugin).getEventManager().preparePartyPreDeleteEvent(party, DeleteCause.LEAVE, null, partyPlayer);
+					((PartiesPlugin) plugin).getEventManager().callEvent(partiesPreDeleteEvent);
 					
-					party.delete();
-					
-					// Calling Post API event
-					IPartyPostDeleteEvent partiesPostDeleteEvent = ((PartiesPlugin) plugin).getEventManager().preparePartyPostDeleteEvent(party.getName(), DeleteCause.LEAVE, null, partyPlayer);
-					((PartiesPlugin) plugin).getEventManager().callEvent(partiesPostDeleteEvent);
-					
-					plugin.getLoggerManager().log(PartiesConstants.DEBUG_CMD_LEAVE_DISBAND
-							.replace("{player}", sender.getName())
-							.replace("{party}", party.getName()), true);
-				} else
-					plugin.getLoggerManager().log(PartiesConstants.DEBUG_API_DELETEEVENT_DENY
-							.replace("{party}", party.getName())
-							.replace("{player}", sender.getName()), true);
+					if (!partiesPreDeleteEvent.isCancelled()) {
+						// Disbanding party
+						sendMessage(sender, partyPlayer, Messages.MAINCMD_LEAVE_LEFT, party);
+						party.broadcastMessage(Messages.MAINCMD_LEAVE_DISBANDED, partyPlayer);
+						
+						party.delete();
+						
+						// Calling Post API event
+						IPartyPostDeleteEvent partiesPostDeleteEvent = ((PartiesPlugin) plugin).getEventManager().preparePartyPostDeleteEvent(party.getName(), DeleteCause.LEAVE, null, partyPlayer);
+						((PartiesPlugin) plugin).getEventManager().callEvent(partiesPostDeleteEvent);
+						
+						plugin.getLoggerManager().logDebug(String.format(PartiesConstants.DEBUG_CMD_LEAVE,
+								partyPlayer.getName(), party.getName(), true), true);
+					} else
+						plugin.getLoggerManager().log(PartiesConstants.DEBUG_API_DELETEEVENT_DENY
+								.replace("{party}", party.getName())
+								.replace("{player}", sender.getName()), true);
+				}
 			} else {
-				party.getMembers().remove(sender.getUUID());
 				party.removeMember(partyPlayer);
 		
 				sendMessage(sender, partyPlayer, Messages.MAINCMD_LEAVE_LEFT, party);
 				party.broadcastMessage(Messages.MAINCMD_LEAVE_BROADCAST, partyPlayer);
 				
-				plugin.getLoggerManager().log(PartiesConstants.DEBUG_CMD_LEAVE
-						.replace("{player}", sender.getName())
-						.replace("{party}", party.getName()), true);
+				plugin.getLoggerManager().logDebug(String.format(PartiesConstants.DEBUG_CMD_LEAVE,
+						partyPlayer.getName(), party.getName(), false), true);
 			}
 			
 			// Calling API event
